move net functionality into pkg/net
Continues the rounds of package separation, this time moving all
network-related functionality into `ghw/pkg/net`
